+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 Lists the changes in the provider.
 
+## Version 0.6.1
+
+- allow dynamic configuration of the provider by introducing a "dynamic_config"
+  provider config flag. This defaults to `false`. When set to `true` then the
+  provider configuration is only validated when actual resources are read or
+  created. This is for specific use cases like AWS deployments where the CML2
+  instance IP is only known after the EC2 instance has been created.
+- bump the gocmlclient to 0.0.18
+
 ## Version 0.6.0
 
 - allow empty node configurations (fixes #50)